pyproject.toml
[project]
name = "mao-thought-and-philosophy"
version = "0.1.0"
description = ""
authors = [
    {name = "yydsp",email = "yydspei@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "importlib-resources (>=6.5.2,<7.0.0)",
    "ebooklib (>=0.20,<0.21)",
    "beautifulsoup4 (>=4.14.2,<5.0.0)",
    "networkx (>=3.6,<4.0)",
    "openai (>=2.8.1,<3.0.0)",
    "colorama (>=0.4.6,<0.5.0)",
    "python-dotenv (>=1.2.1,<2.0.0)"
]

[tool.poetry]
packages = [{include = "mao_thought_and_philosophy", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
